sr(api): move TrainingSummary model in devcomp
