+++ api/src/devcomp/domain/read-models/TrainingSummary.js
